Simple OSC support
This version reads osc messages at a fixed port. The port and the
message are unswitchable. There is nothing in GUI about this. As is,
this in not good enough for a pull request. For that, make the message
and the port editable in the GUI, and make it possible to disable the
GUI.